backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx[brotli]==0.28.1
curl-cffi==0.7.0b4
pydantic==2.11.9
python-dotenv==1.2.1
certifi